refactored server code as well, only view vehicle submissions is left
