+++ server/src/controllers/VehicleController.ts
@@ -1,27 +1,16 @@
 import { AddVehicleToDB, GetVehiclesFromDB } from "../services/VehicleService";
 import { Request, Response } from "express";
-
+import { generatePictureUrls } from "../utils/functions";
 
 export const AddVehicle = async (req: Request, res: Response) => {
   try {
-    // Ensure files are uploaded
     if (!req.files || !Array.isArray(req.files)) {
       res.status(400).json({ msg: "No files uploaded" });
       return;
     }
-
     const { carModel, price, phoneNumber, numPictures, email, userLoggedIn } =
       req.body;
-
-    // Transform file paths to live URLs
-    const pictureUrls = (req.files as Express.Multer.File[]).map((file) => {
-      const baseUrl = `${req.protocol}://${req.get("host")}`;
-      return `${baseUrl}/UserUploads/${file.filename}`;
-    });
-
-    console.log("Add vehicle invoked with req.body: ", req.body);
-    console.log("Transformed pictureUrls: ", pictureUrls);
-
+    const pictureUrls = generatePictureUrls(req);
     await AddVehicleToDB({
       carModel,
       price,
@@ -31,7 +20,6 @@ export const AddVehicle = async (req: Request, res: Response) => {
       email,
       userLoggedIn,
     });
-
     res.status(200).json({ msg: "Vehicle Added" });
   } catch (error) {
     console.error("Error while adding Vehicle: ", error.message || error);
