+++ server/src/middlewares/Image.ts
@@ -1,0 +1,4 @@
+import multer from "multer";
+import { storage } from "../services/ImageService";
+
+export const upload = multer({ storage });
